+++ app/app.ts
@@ -2,12 +2,12 @@ import Application from '@ember/application';
 import Resolver from 'ember-resolver';
 import loadInitializers from 'ember-load-initializers';
 import config from 'houseninja/config/environment';
-import { init } from 'houseninja/lib/sentry';
 import { Intercom } from '@houseninja/capacitor-intercom';
 import isNativePlatform from 'houseninja/utils/is-native-platform';
+import { initializeTelemetry } from 'houseninja/services/telemetry';
 
-// init sentry
-init();
+// telemetry
+initializeTelemetry();
 
 export default class App extends Application {
   constructor(properties: object | undefined) {