+++ app/components/view/work-order/content/approve-payment/component.ts
@@ -17,6 +17,8 @@ import ToastService from 'houseninja/services/toast';
 import WorkOrder from 'houseninja/models/work-order';
 import { ValidationError } from '@ember-data/adapter/error';
 import Invoice from 'houseninja/models/invoice';
+import MetricsService from 'houseninja/services/metrics';
+import SessionService from 'houseninja/services/session';
 
 import type Property from 'houseninja/models/property';
 import type PaymentMethod from 'houseninja/models/payment-method';
@@ -47,6 +49,8 @@ type Args = {
 export default class WorkOrderApprovePaymentViewContentComponent extends Component<Args> {
   @service declare intercom: IntercomService;
   @service declare router: RouterService;
+  @service declare metrics: MetricsService;
+  @service declare session: SessionService;
   @service declare store: StoreService;
   @service declare toast: ToastService;
 
@@ -57,6 +61,11 @@ export default class WorkOrderApprovePaymentViewContentComponent extends Compone
   @tracked creditCard = this.allPaymentMethods.filter(
     (c: PaymentMethod) => c.isDefault
   )[0];
+
+  constructor(owner: unknown, args: Args) {
+    super(owner, args);
+    this.sendMetrics('session');
+  }
 
   actionSheetOptions: ActionSheetButton[] = [
     {
@@ -113,6 +122,7 @@ export default class WorkOrderApprovePaymentViewContentComponent extends Compone
     try {
       await payment.save(); // this will be long running (probably)
       this.paid = true;
+      this.sendMetrics('success');
     } catch (e) {
       if (!this.paid) {
         const hasGenericError =
@@ -126,6 +136,7 @@ export default class WorkOrderApprovePaymentViewContentComponent extends Compone
         }
         this.toggleIsProcessing();
       }
+      this.sendMetrics('error', 'payment');
       captureException(e as Error);
     }
   }
@@ -160,6 +171,9 @@ export default class WorkOrderApprovePaymentViewContentComponent extends Compone
   selectRoute(): void {
     this.isProcessing = false;
     this.args.model.reload();
+    if (!this.isNativePlatform) {
+      this.session.invalidate();
+    }
     this.router.transitionTo(DashboardRoute.Home);
   }
 
@@ -175,7 +189,12 @@ export default class WorkOrderApprovePaymentViewContentComponent extends Compone
     return this.invoice?.formattedTotal;
   }
 
-  // get creditCard() {
-  //   return this.store.peekAll('credit-card').firstObject;
-  // }
+  sendMetrics(event: string, step?: string): void {
+    if (!this.isNativePlatform) {
+      this.metrics.trackEvent({
+        event: `external.payment-approval.${event}`,
+        properties: { step },
+      });
+    }
+  }
 }